requirements.txt
sentence-transformers
langchain_pinecone
langchain
langchain-google-genai
flask
python-dotenv
PyPDF
google
-e .